Add DPMultiHeadAttention
Summary:
MultiheadAttention is one of the key building blocks for NLP models and we want to support it in pytorch-dp.
While conceptually it's reasonably simple and mostly uses transforms  we already support, we don't support the module out of the box. The reason is the way it's implemented in pytorch - it's build on top of one large method in functional package, which leaves us very few opportunities for introspection.

Workaround we use here is to re-implement the module on top of layers we support (`nn.Linear`, `nn.Dropout`). The big downside is that the significant portion of the code is copy-paste from `nn.MultiheadAttention`. On the flipside, however, we get DP support almost for free.
By "almost" here I mean one new module I've added: `SequenceBias`. It encapsulates appending bias item of shape `(embedding_size)` to the end of the sequence. It makes overall per-sample gradient computation trivial.

This diff introduces one core change - parameter for batch_dimension. Normally we expect batch to be the first dimension, and rely heavily on this fact. However, in sequence processing this may not be the case, as input is often of shape `(L, N, E)`, where L - length of the sequence, N - batch size, E - embedding size

I've added new unit-test, to ensure that our custom implementation `DPMultiheadAttention` is equivalent to `nn.MultiheadAttention` and it's safe to replace one with another.
`DPMultiheadAttention` also supports loading from `state_dict` of    `nn.MultiheadAttention`

Reviewed By: Darktex

Differential Revision: D21204110

fbshipit-source-id: fdd9d4ad4143fb97bafa8d5536de4433fb5de1af