Display release build status on Readme
Summary:
the main branch of pytorch-dp depends on the main branches of `torch` and `torchvision`. This may cause build issues on CircleCI at times as we do not load these unstable versions while running the tests.

This commit adds a status badge to the README indicating the health of the release version of pytorch-dp (release version is the build that would get installed via `pip install`)

Reviewed By: sayanghosh

Differential Revision: D22215204

fbshipit-source-id: 376036f9b941ae1c647da98b1745e4d09b10dc41