improve code quality for per_sample_gradient clip
Summary:
The code for per_sample_gradient is a bit complicated, this diff tries to simplify the code a bit
it also hides dynamic clipping into utils.

main contributions:
* per_sample_gradient has not this simple api: `clip_and_accumulate()`, `pre_step()` (I prefer prepare_for_optimizer_step)
where: `pre_step` should be called once before any optimization step, `clip_and_accumulate` should be called at least once
before `pre_step` [it can be called multiple times if we are doing virtual batches.]
* all confusing statistics collecting code is now out of PerSampleClipper, and it now has a call back function that can be set using `set_on_batch_clip_func` the down stream application can set this function to collect as much statistics as they can.
* sample of the above implemented in imagenet.py

* privacey_engine has `step` and `virtual_step` where `step` calls  `clip_and_accumulate()`,  followed by `pre_step()` and ads noise
and  `virtual_step` just calls  `clip_and_accumulate()`.

update : regardless of the clipping strategy (per layer or flat) the noising
can be done using a flat clipping thresholds which is the norm of all the
layer clipping values. This will increase the noise but for now this is
required for Renyie DP to caclulate a tighter bound on the privacy budget.

Reviewed By: Darktex

Differential Revision: D22149569

fbshipit-source-id: 1e71bd86a277ebc8410b88e883ed1c64e7773fcb