Fix for failing DPMultiHeadAttention test.
Summary:
The initialization scheme was changed to uniform normal and threshold for clipping was increased to make it a no-op, so as to test for the actual test criterion of agreement between aggregated per-sample gradients evaluated by our backward hooks, and aggregated gradients produced by vanilla pytorch.

See task T68950370 for details.

Reviewed By: sayanghosh

Differential Revision: D22290951

fbshipit-source-id: b86f11bf23c8375ec6e0c1cedff63926ef699bc4