Test for per sample gradient computation
Summary:
## Overview
This diff introduces new unit tests for per sample gradients.

**IMPORTANT**: This diff introduce broken tests. They indicate a real problem needs addressing. Please only land this diff as the part of the stack, along with fixes for broken tests.

## Per sample gradient test
Prior to this diff we only had test checking if per sample gradient sum up to a correct value. This is not enough, as we also want to make sure individual per sample gradients are correct, because it affects clipping.

Here we take the definition of per sample gradient as a gradient we would get by running a model with a batch with a single sample.
NB: this means that aggregation will depend on the loss reduction

The new test `per_sample_gradient_test.py` first computes per sample gradient for a large batch. Then it takes each row (=one sample) individually runs it through the same model and compares vanilla gradient for this *microbatch* (size=1) with i-th row of per sample gradients
It checks every layer we support (except for DPMultiheadAttention).

## Loss reduction type test
Currently we make an attempt to support two loss reduction types: `mean` and `sum`.

autograd_grad_sample.py:
```
if loss_type == "mean":
    B = backprops * n
else:  # loss_type == 'sum':
    B = backprops
```

The new test checks if we compute per batch and per sample gradients correctly for both loss types by using `nn.L1Loss(reduction)`

## Test resutrs

|    | per batch gradient | per sample gradient
| loss_type == 'mean' | {emoji:2705} OK | {emoji:2705} OK |
| loss_type == 'sum' | {emoji:2705} OK | {emoji:274c} FAIL

What this means in practice is that we clip individual per sample gradients for sum losses too much (fun fact: per batch gradient would also fail if we use `max_norm=999`. It appears not to be big enough, which is a good sign that something is off)

Reviewed By: karthikprasad

Differential Revision: D22116976

fbshipit-source-id: 84b249dbebff486f37ee7298e79b47ac6638ef86