Support sum loss reduction
Summary:
## Overview
This diff introduces full support for losses with `reduction = 'sum'` (i.e. when losses and gradient are summed over the batch, not averaged as we normally do).
As demonstrated in D22116976, we currently make an unsuccessful attempt to support them. It may seem from the code, that we do support them, when we actually don't, which is misleading.
To address this I've made a few changes to make sure we do the right thing for both `mean` and `sum` losses

## Background
As a reminder, here's how we compute gradients in pytorch-dp
1) compute per-sample gradients based on layer's activations and backporpagations
2) clip per sample gradients (let's say to `max_norm=1`)
3) take an average over all per sample gradients in a batch
4) Add noise with `std=1/batch_size`

In order to fully understand what's going in with gradient scaling in pytorch-dp we have to understand this piece of code from `autograd_grad_sample.py:`
```
if loss_type == "mean":
    B = backprops * n
else:  # loss_type == 'sum':
    B = backprops
```

The reason we multiply `backprops` by `n` is that `backprops` are scaled to `1/n` is the first place. With mean loss, each individual sample loss will be averaged to produce final loss, therefore it comes with a `1/n`  factor in an overall loss:

`Loss = 1/n (sample_loss_1 + sample_loss_2 + ... + sample_loss_n)`

In comparison, if the batch is a single sample,

`Loss = sample_loss`

So when you compute derivatives and do backprop with a batch, each sample gradient will be `n` times smaller than it would've been in a single sample batch.

We don't do it for 'sum' losses, because there's no `1/n` multiplier there.

## What's wrong
With the current flow we have 2 bugs:
1) `loss_type` is never 'sum'. We check for it in `autograd_grad_sample.py`, but we never actually set it anywhere and it silently defaults to `mean` (and the `else` branch is never executed)
2) we never check loss reduction type when we aggregate per sample gradients over a batch. We always average them, even though they should be summed up in some cases

Funny enough, those 2 bugs cancel each other out (at least at a first glance) and doesn't affect per-batch gradient. Indeed, fist we multiply gradients by `n` where we shouldn't have, but then we divide by `n` where we shouldn't have.
The only problem it creates in excessive clipping. Because we clip the gradient values after they're multiplied, but before they're divided by `n`, we end up normalizing per sample gradients with a norm `1/n` instead if 1.

NB: this works fine for `mean` gradients, because multiplying by `n` actually brings them back to their original values instead of artificially scaling them up.

## Solution

I've made the following changes:
1) fixed `loss_type` (now called `loss_reduction`) param and passed it down from `PrivacyEngine` to `autograd_grad_sample.py`
2) Changed `grad_sample` accumulation: now we take average or sum up depending on `loss_reduction`
3) Changed the noise addition accordingly. Because the aggregation depends on the loss reduction type, so should the noise. The scale of the noise should be `max_norm/batch_size` for `mean` losses and just `max_norm` for sum losses

Reviewed By: Darktex

Differential Revision: D22116982

fbshipit-source-id: 0572f0852031224245eac2f41c724b816d0fbfbb