remove naked `print`
Summary:
Simple `print` calls are to be avoided in the library code, as it's impossible for a client to configure the logging (level, destination, etc).

We should use either `warnings` or `logging` modules, depending on the usecase.

This is the only occasion currently (in library code, not scripts) where we use print

Reviewed By: sayanghosh

Differential Revision: D23078428

fbshipit-source-id: f1d73dc0a3d50dc3a812c62e344fe6239192488e