Add GradSampleModule (#56)
Summary:
Pull Request resolved: https://github.com/pytorch/opacus/pull/56

Having grad samples is a core abstraction in DP-SGD. The way we currently handle grad samples is a bit magical with hooks, global variables and all. This introduces a more logical reorganization of our code by promoting grad sample hooks to be not just that the privacy engine attaches at some point, but a core abstraction and concept, in line with the theory.

Differential Revision: D22991710

fbshipit-source-id: a85af91f2b9acc461a561febb9a7cc4fc8382eab