Fix notebook rendering
Summary:
The `nbconvert` library - used to convert notebooks to html - had a major release a week ago and changed the default HTML DOM structure. https://nbconvert.readthedocs.io/en/latest/changelog.html#significant-changes

This resulted in tutroial pages not showing up on the website.

{F336116311}

This commit explicitly specifies the desired template to be "classic" to address the issue

This commit also makes changes to the script so that we catch such issues through CircleCI tests.

Reviewed By: ashkan-software

Differential Revision: D23747980

fbshipit-source-id: 3d4265a1a6141a7d535b30dbc17509faf939b7d4