Make Opacus Python 3.6.9 compatible (#63)
Summary:
Pull Request resolved: https://github.com/pytorch/opacus/pull/63

We recently changed the minimal supported version of Python to 3.6 to be Google Colab friendly. We should, therefore, also ensure our tests and baselines are run on this version.

We also run unittests on Py 3.7 and 3.8.
The unit tests on nightly are only run on 3.8 as we only concern ourselves with the latest versions for that.

Differential Revision: D23796762

fbshipit-source-id: d5f8fffc1f9302105120c183388fd25dcbaf0e31