Make Opacus Python 3.6.9 compatible (#63)
Summary:
Pull Request resolved: https://github.com/pytorch/opacus/pull/63

We recently changed the minimal supported version of Python to 3.6 to be Google Colab friendly. We should, therefore, also ensure our tests and baselines are run on this version.

We also run unit tests on Py 3.7 and 3.8.
The unit tests on nightly are only run on 3.8 as we only concern ourselves with the latest versions for that.

Reviewed By: Darktex

Differential Revision: D23796762

fbshipit-source-id: feabf38734f14e59e7c37e1d924b9190f7ea56dc