Fix batch construction and privacy engine
Summary:
In this PR, we fixed the arguments of the privacy engine in order
to be able to use the correct batch construction. We also update
the batch construction used in all examples and tutorials.

The privacy budget evaluation implemented in Opacus follows [1].
This evaluation is based on the Sampled Gaussian Mechanism. As a
consequence, this analysis assumes that to build one batch, each
sample in the dataset is selected with probability q. In addition,
each batch is built independently from the other batches. Using
another batch construction will break this assumption and the
privacy budget will then be incorrectly evaluated. In particular,
as pointed out in [2], one cannot sample constant-sized batches
through random dataset shuffling at each epoch and use the privacy
accounting method from [1].

[1] Renyi Differential Privacy of the Sampled Gaussian Mechanism
[2] Improving deep learning with differential privacy using gradient
encoding and denoising