Disable failing DPMultiheadAttention test
Summary:
## Background
In D25318869 (https://github.com/pytorch/opacus/commit/46a81f99c22cdd40ef20636b7d53963286ba356a) (PR #120) Davide introduced more tests for `dp_layers` package. One of those tests was failing, uncovering a real issue we need to fix.
The underlying problem itself is not high-pri and so it was deemed to be a good first task for the new contributors. Davide then opened an issue on github (https://github.com/pytorch/opacus/issues/123) with a detailed guide for anyone who wants to help. So far there's been no takers, and we wait.

## Issue
Given that the issue is not high-pri and we're aware of it, I don't think we should be using unit tests as a task tracker.
We shouldn't treat red CircleCI as a norm, as it could potentially mask real issues.

## Solution
This diff disables the test in question from running on CircleCI.
I'll comment on the github issue to ask future author of the fix to re-enable the test after the issue is fixed.

Reviewed By: karthikprasad

Differential Revision: D26277444

fbshipit-source-id: 9d5888c0f5f5e5cb88952370f13f305b143ee825