Fix batch construction and privacy engine (#128)
Summary:
## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Docs change / refactoring / dependency upgrade

## Motivation and Context / Related issue

In this PR, we fixed the arguments of the privacy engine in order
to be able to use the correct batch construction. We also update
the batch construction used in all examples and tutorials.

The privacy budget evaluation implemented in Opacus follows [1].
This evaluation is based on the Sampled Gaussian Mechanism. As a
consequence, this analysis assumes that to build one batch, each
sample in the dataset is selected with probability q. In addition,
each batch is built independently from the other batches. Using
another batch construction will break this assumption and the
privacy budget will then be incorrectly evaluated. In particular,
as pointed out in [2], one cannot sample constant-sized batches
through random dataset shuffling at each epoch and use the privacy
accounting method from [1].

[1] Rényi Differential Privacy of the Sampled Gaussian Mechanism
[2] Improving deep learning with differential privacy using gradient
encoding and denoising

## How Has This Been Tested (if it applies)

This commit has been tested on examples and tutorials.

## Checklist

- [x] The documentation is up-to-date with the changes I made.
- [x] I have read the **CONTRIBUTING** document and completed the CLA (see **CONTRIBUTING**).
- [x] All tests passed, and additional code has been covered with new tests.

All tests except one passed. This failed test was already broken before this commit.

Pull Request resolved: https://github.com/pytorch/opacus/pull/128

Reviewed By: karthikprasad

Differential Revision: D26559465

Pulled By: Darktex

fbshipit-source-id: 342d913fc9255e3f420856b40962d240fd4c02ec