Switch MNIST integration test to latest torchvision (#156)
Summary:
Pull Request resolved: https://github.com/pytorch/opacus/pull/156

yann.lecun.com is having availability issues, which leads to our integration tests failing.

torchvision addressed the issue recently - the change haven't yet made it to the release, but it works if you install from sources.

I suggest we temporary add github installation into CircleCI config for now, and then move to the latest torchvision as soon as it's released

Reviewed By: karthikprasad

Differential Revision: D27208798

fbshipit-source-id: 84a59a00500d97e400f62904c01614aef1c0640f