Downgrade to torch==1.8.1 for tests (#188)
Summary:
Pull Request resolved: https://github.com/pytorch/opacus/pull/188

~~Pip install on dev_requirements started to time out recently.
One possible reason is dependency resolver, which has almost no constraints and takes too long to find satisfying versions.~~

```
INFO: This is taking longer than usual. You might need to provide the dependency resolver with stricter constraints to reduce runtime.
```

~~Let's see if reducing the search space helps.~~

~~NB: all package versions are what was the latest available at the end of 2020~~

Ok, I was wrong

The actual problem was pytorch 1.9 release.
We have no problem supporting it, but latest `torchcsprng` has a pegged dependency on 1.8.

So what happens is
1) we install regular `requirements.txt` with `torch==1.9.0`
2) we try to install `torchcsprng`, which requires 1.8, can't and fail.

The strange timeout behaviour is down to the [new pip dependency resolver](https://pyfound.blogspot.com/2020/11/pip-20-3-new-resolver.html) - instead of flagging the conflict it tries to traverse the graph and find the non-conflicting combination. Clearly, it's impossible, but brave and tenacious dependency resolver timeouts, instead of telling what's wrong

(fun fact: debugging this involved running `pip install --use-deprecated=legacy-resolver`, which is not a great sign for a new feature)

PS:
I thought that the correct way to address this was to install `requirements.txt` and `dev_requirements.txt` at the same time, so the pip would figure out himself that the only option is 1.8. Unfortunately, it doesn't - it installs 1.9 and then fails. Not so smart after all

PS2:
Non-dev dependency stays as `torch>=1.3`, so people will be able to use opacus with torch 1.9

Reviewed By: alexandresablayrolles

Differential Revision: D29265512

fbshipit-source-id: 868500a1c79e781b26081e6f9d44517fb03abcd3