Remove yarn.lock and update the docstrings for website
Summary:
Recently, we had several issues regarding vulnerabilities in `yarn.lock`. mironov suggested maybe we can get rid of `yarn.lock` altogether, if it's an auto-generated file. Turns out he was right, and we can get rid of it! (`package.json`) will create it.

In addition to that change, this diff also adds the GradSampleModule to the docstrings on the website, and updates its docstrings

Reviewed By: ffuuugor

Differential Revision: D29277571

fbshipit-source-id: c0d1b95a09fb7d273bcbe7ec194d9b5a0d1f9066