imdb integration test fix
Summary:
Recently IMDB integration test started failing.
Th culprit is this line:
```
load_dataset("imdb", cache_dir=args.data_root)
```

The problem is (new?) beaviour of `load_dataset`, or in particular - how it treats the first argument. It could mean either the name of the dataset (in which case it'll download it) or a path (in which case it'll try to load the dataset from the path given).
The latter interpretation has priority if the path exists - which it does in our integration test setup - we create a temporary folder with the unfortunate name "imdb".

Solution: we put all temporary files into `runs/` subdirectory.

Reviewed By: karthikprasad

Differential Revision: D31116926

fbshipit-source-id: d35c26b8eb88d8a10108db8885cc2b6c5046c0dc