imdb integration test fix (#208)
Summary:
Pull Request resolved: https://github.com/pytorch/opacus/pull/208

Recently IMDB integration test started failing.
Th culprit is this line:
```
load_dataset("imdb", cache_dir=args.data_root)
```

The problem is (new?) beaviour of `load_dataset`, or in particular - how it treats the first argument. It could mean either the name of the dataset (in which case it'll download it) or a path (in which case it'll try to load the dataset from the path given).
The latter interpretation has priority if the path exists - which it does in our integration test setup - we create a temporary folder with the unfortunate name "imdb".

Solution: we put all temporary files into `runs/` subdirectory.

Reviewed By: karthikprasad, ashkan-software

Differential Revision: D31116926

fbshipit-source-id: d6591a874e36bea89432bb805a1f7573b30110d9