Support of GRU and RNN. Refactored LSTM implementation. (#222)
Summary:
## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [x] Docs change / refactoring / dependency upgrade

## Motivation and Context / Related issue

The original issue is to implement GRU support for opacus https://github.com/pytorch/opacus/issues/206.
This requires creating a drop-in replacement for `nn.GRU` similar to `DPLSTM` (a replacement for `nn.LSTM`). Both LSTM and GRU (as well as simple RNN) use `nn.Linear` as the only parametrized building block, so we will reuse the special version of it equipped with accumulating grad sampler (called `LSTMLinear` in the code).

## Changes

The idea of this PR is to create a set of classes with exactly the same interface as in [`torch.nn.modules.rnn`](https://github.com/pytorch/pytorch/blob/master/torch/nn/modules/rnn.py).

1. LSTMLinear (the basic building block of DPLSTM) renamed to more generic `RNNLinear`
2. Added `DPRNNCellBase` class and DP**RNN**Cell, DP**GRU**Cell, DP**LSTM**Cell derived from it.
3. Added `DPRNNBase` class and DP**RNN**, DP**GRU**, DP**LSTM** derived from it. Supported by grad sampler.
4. In `char-lstm-classification.py` example added flag `--mode {rnn,gru,lstm}` to choose between different models
5. `ParamRenamedModule` is replaced with `RenameParamsMixin` trait
6. `compute_seq_lengths` is moved to `opacus.utils.packed_sequences` as it was double implemented, it is useful and has clear interface.

All classes implemented in `dp_rnn` have identical interface and `state_dict` with those from `torch.nn.rnn`.

The only thing that is not implemented: the case with `proj_size > 0` (it wasn't implemented before). It can be a new issue for external contributors which is much smaller and easier for the newcomers.

## How Has This Been Tested

- Updated tests `dp_layers/dp_rnn_test.py` and `grad_samples/dp_rnn_test.py` to cover the new features.
- Pytest passes
- Modified `examples/char-lstm-classification.py` to switch between RNN/GRU/LSTM. Run all models and checked learning curves.

Lints:
- Applied `black` reformatting
- Applied `isort`
- Checked flake8

## Checklist

- [x] The documentation is up-to-date with the changes I made.
- [x] I have read the **CONTRIBUTING** document and completed the CLA (see **CONTRIBUTING**).
- [x] All tests passed, and additional code has been covered with new tests.

Pull Request resolved: https://github.com/pytorch/opacus/pull/222

Reviewed By: ffuuugor

Differential Revision: D31681985

Pulled By: romovpa

fbshipit-source-id: e29d42c718345e69cfa32b8284b71b8dc7c31446