Add support for GDP accounting in get_noise_multiplier (#362)
Summary:
Pull Request resolved: https://github.com/pytorch/opacus/pull/362

The PR adds support for GDP accounting in get_noise_multiplier.

This is done by making the GDP accountant work more like the RDP accountant: Keep the list of parameters in `self.steps` (as opposed to having `self.steps` being an integer as before). This change in behavior, however, does not affect the public interface of the GDP accountant.

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Docs change / refactoring / dependency upgrade

## Motivation and Context / Related issue

This PR solves a TODO in code. The problem is that the `get_noise_multiplier` function currently cannot leverage the GDP accounting mechanism.

## How Has This Been Tested (if it applies)
A respective unit test was added in `opacus/tests/accountants_test.py`

## Checklist

- [x] The documentation is up-to-date with the changes I made.
- [x] I have read the **CONTRIBUTING** document and completed the CLA (see **CONTRIBUTING**).
- [x] All tests passed, and additional code has been covered with new tests.

Pull Request resolved: https://github.com/pytorch/opacus/pull/303

Reviewed By: alexandresablayrolles

Differential Revision: D32933812

Pulled By: ashkan-software

fbshipit-source-id: eb8d29af12f0908a38f8269cc29ff96140c60b1b