Extract the input projections to a dedicated module.
Summary:
Three linear layers that were used to transform the inputs query, key and value respectively, were extracted to a dedicated module InputProjection.

In addition to the extraction refactoring, the nn.Linear modules used for the transformations were replaced with a bespoke operator parameterized by 4 tensors. That same operator is used in the reference module nn.MultiheadAttention. The modification will allow to match the states of nn.MultiheadAttention and opacus.DPMultiheadAttention.

Lastly, a custom gradient sampler was defined for the new module. This preserves the gradient sampling compatibility of the parent module - DPMultiheadAttention.

Differential Revision: https://www.internalfb.com/diff/D34850578?entry_point=27

fbshipit-source-id: 9b8cc1a6f969c74551650c33bf18b3ddf241d226