Define a module for packed input projection
Summary:
Added a new module that projects the three inputs of a multihead attention module. In contrast to the unpacked version, this module uses an operator that works with only two tensors.

The operator employed is the same one that is used by nn.MultiheadAttention in the case when the shapes of all three inputs match.

Differential Revision: D34850577

fbshipit-source-id: 70c7bd7fb02593456d6adbc607d06b6b197ce85a