Support tied parameters (#417)
Summary:
## Solution

As reported in https://github.com/pytorch/opacus/issues/413, some models (including huggingface GPT-2) use tied parameters - i.e. sharing the same parameter between modules by simply using the same Parameter object.
Opacus gets confused - it treats the second set of grad_samples as a new batch and accumulates them by stacking, which is obviously wrong. Conceptually this is not very different from RNNs, and the gradients should be simply added up.

Natural way to support this is to tweak the condition on which we distinguish between recurrent calls and new batch. Currently we look at module-level activations - if we've run out of the activations, it means that the recurrent calls are over and the next grad sample will be from a new batch. This obviously doesn't work for shared parameters - as the condition is module-level, but the parameter belong to multiple modules.

What we can do instead is to have parameter-level counters (`p._forward_counter`). This counter would increase on forward pass and decrease on backward. When it gets to zero, in would indicate that the latest backward pass is fully unrolled and we should expect new forward pass with the new data. If I understand correctly, we do not and should not allow multiple forwards or backwards in a row.

## Tests

I've added a new unit test for this and also checked Colab from the bug report - all seem to be working.

Pull Request resolved: https://github.com/pytorch/opacus/pull/417

Reviewed By: karthikprasad

Differential Revision: D35843923

Pulled By: ffuuugor

fbshipit-source-id: 09370a562b1bd22a0825fb8ce655c51784036270