zero_grad() between forward and backward passes generates error message
Summary:
In case of Poisson sampling, we must check that all `grad_sample` attributes are set to `None` (since we forbit accumulation with Poisson sampling).

This was previously done in a forward hook, resulting in an error message being thrown inadequately, in particular when `zero_grad()` is called between the forward and backward steps (see task for more context).

Hence, we instead register a backward hook at the `GradSampleModule` level (to ensure it's the first hook to be called, before the hooks that populate the `grad_sample` attribute). Updated the docstring accordingly with more details.

Differential Revision: D36067469

fbshipit-source-id: 583b1865e1d1e30b50f695fa3e7dc1dbb528233d