Update faq.md
Added `accountant` in line 113. PrivacyEngine has no function `get_privacy_spent(delta)`. Also, the call requires delta to be passed as a named argument.

Updated line 111 to note how a custom alpha list can be provided. There is no mention of `alphas` in the `privacy_engine` documentation, and I had to follow the code to figure it out.