reduce logging severity for set_to_none (#471)
Summary:
We want to warn users of an unexpected behaviour with `set_to_none` flag. Normally, both `nn.Module` and `Optimizer` let clients choose whether they want to remove the `.grad` attribute altogether or just set it to None.
We, on the other hand, don't want to remove the attributes - it's more convenient to assume the `.grad_sample` attribute is always present. It's not an absolute requirement, but we did that historically and I don't see a case for changing it now.

However, default value for `set_to_none` is False, meaning most users are getting annoying logging notifications on every training step.

Pull Request resolved: https://github.com/pytorch/opacus/pull/471

Reviewed By: karthikprasad

Differential Revision: D38741747

Pulled By: ffuuugor

fbshipit-source-id: e77c4ec90c7aa70866d2f073034bd2fd2c27c476