No-Op GradSampleModule (#492)
Summary:
TL;DR: Adding a No-Op GradSampleModule in case the grad samples are computed by functorch. The CIFAR10 example has been updated to show a typical use-case for that.

The neat thing about functorch is that it directly gives the per-sample gradients with a couple of lines of code. These per-sample gradients are then manually given to `p.grad_sample` by the end-user.

Pull Request resolved: https://github.com/pytorch/opacus/pull/492

Reviewed By: ffuuugor

Differential Revision: D39204008

Pulled By: alexandresablayrolles

fbshipit-source-id: 22036e6c941522bba7749ef46f97d54f6ee8c551