Fixing Ghost Clipping with Batch Memory Manager (#707)
Summary:

Ghost Clipping with Batch memory manager had an error, resulting in major accuracy loss. The issue was in the accumulate function, the command `p.summed_grad +=p.grad`, wasn't working as expected, since `p.grad` is modified in every iteration. The fix is to copy it and modify in-place.

Reviewed By: HuanyuZhang

Differential Revision: D67778159