Fix IAccountant.mechanism accessed as a property (#778)
Summary:
## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Docs change / refactoring / dependency upgrade

## Motivation and Context / Related issue

`Accountant.mechanism` is used as a class property in IAccountant state_dict methods, while it is an overloadable function in the code. It breaks state_dict functionality both ways.

## How Has This Been Tested (if it applies)

This modification has been used to implement a checkpointer for use in with Transformers library. It fails without it and works just fine with the fix.

Test was broken as well.

## Checklist

- [x] The documentation is up-to-date with the changes I made.
- [x] I have read the **CONTRIBUTING** document and completed the CLA (see **CONTRIBUTING**).
- [x] All tests passed, and additional code has been covered with new tests.

Pull Request resolved: https://github.com/pytorch/opacus/pull/778

Reviewed By: EnayatUllah

Differential Revision: D79565816

Pulled By: iden-kalemaj

fbshipit-source-id: 29ea4ee069d4f98b01fa809a20bafde83e9f02d2